requirements.txt
pytest-asyncio==0.14.0
black==19.10b0
flake8==3.8.3
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
mypy
wheel
